BetterFps FastBeacon Compatibility
Puts the Beacon below y0 fix into the selectable mixins and tries to discover BetterFps and its FastBeacon setting. If that setting is found active, force disables the Beacon below y0 fix to avoid a crash.
Disable FastBeacon of BetterFps in total to activate, or use a CC-aware BetterFps version instead.