Fix scheduled block ticks.
Fixes issue with water, redstone and other blocks not always continuing
after reloading world.
Fixes incorrect redstone behavior (for example repeater based clock
not working).

Instead of moving scheduled ticks to Cubes, keep them in WorldServer
and replace the collections that would normally be too slow to filter
for all of the cubes with custom implementation.