Fix a crash when trying to load a tile entity outside of the cube it's in.
This crashes with StackOverflowError in cubic chunks because of a "hack"
to actually be able to do it through chunk, using caching.
After the first TE with the wrong position is loaded, a different cube
gets cached so the next TE add will try to load the original cube again.
Which will still contain the out of cube TE, and trigger loading itself
again, until StackOverflowError.