Fix cubes not being sent to client with certain corner case settings and chunk states
When max cube generation per tick was set to 1, it would get stuck trying to generate lighting for a cube that didn't have it's neighbors yet, and it would never reach a state where it could be updated because only the same one cube would be processed per tick.

There are multiple fixes included:
 * properly consider whether a cube has been actually generated. Should no longer count a cube as generated when lighting hasn't been finished in it
 * Give up trying to generate lighting on third failed attempt and consider it done anyway

Also centralized checks for the cube generation state to avoid duplicating the same code with slightly different logic.